testes/2-transformacao_validacao/requirements.txt
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
